Rename and refactor OrderWorkflowServiceTest - Previous: OrderWorkflowControllerTest - Rename the method: GetOrderWorkflow -> GetOrderWorkflow_ByOrderId_Should_Return_Workflow - Use AutoFixture to minimize effort of setting up test scenario - Use FluentAssertions to make the assertions look beautiful, natural and most importantly, extremely readable
