--- VirtoCommerce.OrderModule.Test/OrderWorkflowControllerTest.cs
@@ -1,64 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Configuration;
-using System.Data.Common;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using CacheManager.Core;
-using Microsoft.Practices.Unity;
-using Moq;
-using VirtoCommerce.Domain.Customer.Model;
-using VirtoCommerce.OrderModule.Core.Models;
-using VirtoCommerce.OrderModule.Core.Services;
-using VirtoCommerce.OrderModule.Data.Model;
-using VirtoCommerce.OrderModule.Data.Repositories;
-using VirtoCommerce.OrderModule.Data.Services;
-using VirtoCommerce.Platform.Core.Assets;
-using VirtoCommerce.Platform.Core.Common;
-using VirtoCommerce.Platform.Data.Infrastructure;
-using VirtoCommerce.Platform.Data.Infrastructure.Interceptors;
-using VirtoCommerce.Platform.Testing.Bases;
-using Xunit;
-using Moq.Protected;
-
-
-
-namespace VirtoCommerce.OrderModule.Test
-{
-    [CLSCompliant(false)]
-    [Trait("OrderWorkflow", "OWF")]
-    public class OrderWorkflowControllerTest : TestBase
-    {
-
-        [Fact]
-        public void GetOrderWorkflowTest()
-        {
-            //arrange
-            var orderRepositoryMock = new Mock<IOrderRepository>();
-            string workflowId = Guid.NewGuid().ToString();
-            string orderId = Guid.NewGuid().ToString();
-            List<OrderWorkflowEntity> OrderWorkflowTestData = new List<OrderWorkflowEntity>
-            {
-                new OrderWorkflowEntity
-                {
-                    WorkflowId = workflowId,
-                    CreatedDate = DateTime.Now,
-                    ModifiedDate = DateTime.Now,
-                    OrderId = orderId,
-                    Id = Guid.NewGuid().ToString(),
-
-                },
-            };
-            orderRepositoryMock.Setup(x => x.OrderWorkflows).Returns(OrderWorkflowTestData.AsQueryable());
-
-            //act            
-            var orderService = new OrderWorkflowService(orderRepositoryMock.Object);
-            var result = orderService.GetOrderWorkflow(orderId);
-
-            //assert
-            Assert.Equal(workflowId, result.WorkflowId);
-        }
-    }
-}
