+++ VirtoCommerce.OrderModule.Test/VirtoCommerce.OrderModule.Test.csproj
@@ -166,7 +166,7 @@
     <Compile Include="MigrationScenarios.cs" />
     <Compile Include="OrderInventoryAdjustmentTests.cs" />
     <Compile Include="OrderTotalsCalculationTest.cs" />
-    <Compile Include="OrderWorkflowControllerTest.cs" />
+    <Compile Include="OrderWorkflowServiceTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="CRUDScenarios.cs" />
     <Compile Include="SequencesScenarios.cs" />
